requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# GPU acceleration (optional but recommended)
cupy-cuda12x>=12.0.0  # For CUDA 12.x - adjust version for your CUDA installation
# cupy-cuda11x>=11.0.0  # Alternative for CUDA 11.x

# Machine Learning & Imputation
scikit-learn>=1.3.0
statsmodels>=0.14.0
cuml-cu12>=24.0.0  # GPU-accelerated sklearn (optional)

# Genetic Programming
deap>=1.4.0

# Advanced Imputers
# fancyimpute>=0.7.0  # Optional - can be problematic to install
# missingpy>=0.2.0    # Optional - MissForest alternative

# Statistical Tests
scikit-posthocs>=0.7.0

# Plotting & Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Hyperparameter Optimization
optuna>=3.0.0

# UCI ML Repository
ucimlrepo>=0.0.3

# Boosting
xgboost>=2.0.0

# Configuration & Utils
pyyaml>=6.0
joblib>=1.3.0
tqdm>=4.65.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.2.0
